+++ src/resources/submission/index.jsx
@@ -1,0 +1,2 @@
+export { SubmissionList } from './list.jsx';
+export { SubmissionShow } from './show.jsx';